fix(AV-1667): Guide page style fixes
Remove top padding from arrowbox so it is placed under the main menu.
Change the font weight of the text in the arrowbox to 600.

Change color of the hover styling on the menu items in the drop downs.
The entries in the drop down menus move a few pixels upwards when hovering over them,
remove the bottom border to fix this. This has a side effect on the main menu when no page is active
due to the minimum size of the main menu, change the main menu min-height to 67px.

Refs AV-1667